pyproject.toml
[project]
name = "who-killed-jfk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.13.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]
